Fixed movement path preview bug
There was an issue if you hovered into the vantage point where it would mark the path for destruction but then go to check if it exists in the same frame. And it still did! Cause Unity's Destroy waits until the end of the current frame. But they do have DestroyImmediate so 🍤